fix: changed level type

+++ src/main/java/com/greenteam/schoolmanager/dto/userSkill/UserSkillEntityCreationPayload.java
@@ -13,7 +13,7 @@ public class UserSkillEntityCreationPayload {
 
     @NotNull
     @Min(1) @Max(5)
-    private Integer level;
+    private Float level;
 
     @NotNull
     @PositiveOrZero
